Middleware

+++ routes/web.php
@@ -20,17 +20,20 @@ Route::get('/', function () {
 Auth::routes();
 
 Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
-//Rutas de Especialidades
-Route::get('/especialidades', [App\Http\Controllers\SpecialityController::class, 'index']);
-Route::get('/especialidades/create', [App\Http\Controllers\SpecialityController::class, 'create']);
-Route::get('/especialidades/{speciality}/edit', [App\Http\Controllers\SpecialityController::class, 'edit']);
-Route::post('/especialidades', [App\Http\Controllers\SpecialityController::class, 'sendData']);
 
-Route::put('/especialidades/{speciality}', [App\Http\Controllers\SpecialityController::class, 'update']);
-Route::delete('/especialidades/{speciality}', [App\Http\Controllers\SpecialityController::class, 'destroy']);
+Route::middleware(['auth', 'admin'])->group(function () {
+   //Rutas de Especialidades
+Route::get('/especialidades', [App\Http\Controllers\admin\SpecialityController::class, 'index']);
+Route::get('/especialidades/create', [App\Http\Controllers\admin\SpecialityController::class, 'create']);
+Route::get('/especialidades/{speciality}/edit', [App\Http\Controllers\admin\SpecialityController::class, 'edit']);
+Route::post('/especialidades', [App\Http\Controllers\admin\SpecialityController::class, 'sendData']);
+
+Route::put('/especialidades/{speciality}', [App\Http\Controllers\admin\SpecialityController::class, 'update']);
+Route::delete('/especialidades/{speciality}', [App\Http\Controllers\admin\SpecialityController::class, 'destroy']);
 
 //Rutas de Medicos
-Route::resource('medicos','App\Http\Controllers\DoctorController');
+Route::resource('medicos','App\Http\Controllers\admin\DoctorController');
 
 // Rutas Pacientes
-Route::resource('pacientes','App\Http\Controllers\PatientController');
+Route::resource('pacientes','App\Http\Controllers\admin\PatientController');
+});
